Add `babel-preset-env` to allow smaller builds
Can disable transforms/plugins and buildIns (polyfills) from a config
that is easily changeable.

Currently the largest and rarely used features are disabled, and can
be turned on when needed.

Also, there is no size difference between IE9 and IE11 in the
browserslist version array.

Added a separate polyfills.js file that is added as the first webpack
entry that will contains all polyfill related code, that is now also
pulling in state 3 and 4 features from core-js.