Update structure to allow for custom builds
Previously all scripts where included separately, so they were
disconnected from the templates they are used in. With this separation
it's not possible to create a custom build containing only a few
of the blocks, without manually specifying each and every single script.

To address this, the hbs-style-loader is upgraded to also allow script
includes. Besides the require, it will also register the class
globally, sot can be constructed for all elements in the DOM.

When running in development mode, it will also include the HOT snippet
to allow live changes.

Because of this change, all script and templates were outputted as a
single bundle. Since the templates are only needed to generate the html
pages, we don't want them together with the scripts.

To address that issue, the loader has an option to completely remove
the original contents of the hbs file, and only strip out the requires.
With that, it's possible to control what is in the output; scripts,
styles, templates, or a combination of them.

The webpack build has ben updated to dynamically add entries to the
config based on the json files and the blocks specified in them.

To improve things, webpack is also outputting a common.js file with
all node_modules, and a common.css that include all global styles.
These common.js/css are now included in the build html files.

The build script now also outputs 'standalone' html files that include
the custom css and js files.

Outputting css is now done from the `.dist` webpack config, and the
`.partials` webpack config now only outputs the partials to generate the
html files.

re #15