Add util function to clean up a HTML element before replacing it
Remove all instances bound to this html element or its children.
Finds all elements with a data-component attribute, and disposes and
removes the created component instance for that element.

You should call this function before removing/replacing any piece of
HTML that has components attached to it (e.g. when calling
`initComponents` on replaced HTML)