Add utils and documentation for working with dynamic data
Add documentation that describes some examples and best practices for
working with dynamic data (fetch from the backend).

Add the `updateElement` util to update the HTML content of an element,
including cleanup and initializing the new components. This can be
useful when you retrieved new HTML from the backend and need to replace
a section of the page.

Add the `initTextBinding` util to set up a knockout binding to an
element and include its content as the initial value for the observable.

Add the `initListBinding` util to set up a knockout binding for a list
that renders using a knockout template. It has the option to extract
data from the already rendered items (or you can pass them manually),
so they can be re-rendered client-side.