Add Knockout component and Redux integration
Add knockout-loader, that allows you to import styles and a ViewModel
script from a knockout component template, which registers itself as
a knockout component.

Add a `knockout-redux-connect` function that connects knockout with
redux, converting the returned redux state to observables, and updating
them when the store changes.

Added documentation for the above; how to set up a knockout and
redux 'block' in your project.

re #27