Let lint-staged only run on staged files
The 'src' paths from the normal lint-scripts were also being included when run via lint-staged, which cause the complete project to be linted on every commit. Copying over the commands to the lint-staged config and removing the paths will fix this issue.

Please make sure to apply changes on both locations when required.