Add Freemarker (ftl) support to muban
* Add a dev and compile task to compile FTL templates to HTML
* Duplicate current hbs templates into FTL variants
* Add new build-ftl script, and additional build targets in the build script
* Add a new preview-server with live-reload

The ftl templates will be built next to the normal HTML templates, with
the `-ftl.html` postfix. During dev, the dev:code will be called in with
watch mode, and a separate watcher will trigger ftl template rebuilds.
On any change, the live-reload server will be triggered so the page is
refreshed.