Enable prefixIDs svgo plugin
The cleanupIDs plugin in svgo minifies IDs, it does this for each file. The scope for inline svg elements is the document in which they're placed, this means that the id attribute value should be unique to make sure that the correct svg definition is used.

When two svg files have id attributes the values will collide, invalid HTML is generated when both files are inlined. This can cause issues in the svg files that have references to the definitions because it might reference the incorrect element.